requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.7
django-admin==2.0.2
django-allauth==0.58.2
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
excel-base==1.0.4
idna==3.4
isoweek==1.3.3
oauthlib==3.2.2
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.4
TimeConvert==3.0.13
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.0.7
xlwt==1.3.0
